+++ src/zepben/evolve/services/network/network_service.py
@@ -11,9 +11,7 @@ import itertools
 import logging
 from enum import Enum
 from pathlib import Path
-from typing import TYPE_CHECKING, Dict, List, Union, Iterable, Optional, Generator, Set
-
-from zepben.evolve.util import ngen
+from typing import TYPE_CHECKING, Dict, List, Union, Iterable, Optional, Set
 
 from zepben.evolve.model.cim.iec61970.base.auxiliaryequipment.auxiliary_equipment import AuxiliaryEquipment
 from zepben.evolve.model.cim.iec61970.infiec61970.feeder.lv_feeder import LvFeeder